+++ irs-api/src/main/java/org/eclipse/tractusx/irs/connector/job/PersistentJobStore.java
@@ -37,6 +37,7 @@ import org.eclipse.tractusx.irs.component.enums.JobState;
 import org.eclipse.tractusx.irs.data.JsonParseException;
 import org.eclipse.tractusx.irs.common.persistence.BlobPersistence;
 import org.eclipse.tractusx.irs.common.persistence.BlobPersistenceException;
+import org.eclipse.tractusx.irs.services.DataIntegrityService;
 import org.eclipse.tractusx.irs.services.MeterRegistryService;
 import org.eclipse.tractusx.irs.util.JsonUtil;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -61,8 +62,8 @@ public class PersistentJobStore extends BaseJobStore {
     private final MeterRegistryService meterService;
 
     public PersistentJobStore(@Qualifier(JOB_BLOB_PERSISTENCE) final BlobPersistence blobStore,
-            final MeterRegistryService meterService) {
-        super();
+            final MeterRegistryService meterService, final DataIntegrityService dataIntegrityService) {
+        super(dataIntegrityService);
         this.blobStore = blobStore;
         this.meterService = meterService;
     }