+++ irs-api/src/main/java/org/eclipse/tractusx/irs/connector/job/RecursiveJobHandler.java
@@ -24,6 +24,8 @@ package org.eclipse.tractusx.irs.connector.job;
 
 import java.util.stream.Stream;
 
+import org.eclipse.tractusx.irs.aaswrapper.job.ItemContainer;
+
 /**
  * Interface for extensions to provide the logic to build jobs with
  * custom logic to run multiple transfers.
@@ -51,6 +53,7 @@ public interface RecursiveJobHandler<T extends DataRequest, P extends TransferPr
      * Called when all transfers in the job have completed.
      *
      * @param job job definition.
+     * @return asd
      */
-    void complete(MultiTransferJob job);
+    ItemContainer complete(MultiTransferJob job);
 }