+++ irs-api/src/test/java/org/eclipse/tractusx/irs/connector/job/InMemoryJobStoreTest.java
@@ -27,14 +27,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.eclipse.tractusx.irs.aaswrapper.job.ItemContainer;
 import org.eclipse.tractusx.irs.component.Job;
 import org.eclipse.tractusx.irs.component.JobErrorDetails;
 import org.eclipse.tractusx.irs.component.enums.JobState;
+import org.eclipse.tractusx.irs.services.DataIntegrityService;
 import org.eclipse.tractusx.irs.util.TestMother;
 import net.datafaker.Faker;
 import org.assertj.core.api.SoftAssertions;
@@ -42,7 +46,7 @@ import org.junit.jupiter.api.Test;
 
 class InMemoryJobStoreTest {
     final int TTL_IN_HOUR_SECONDS = 3600;
-    InMemoryJobStore sut = new InMemoryJobStore();
+    InMemoryJobStore sut = new InMemoryJobStore(new DataIntegrityService(""));
     Faker faker = new Faker();
     TestMother generate = new TestMother();
     MultiTransferJob job = generate.job(JobState.UNSAVED);
@@ -183,7 +187,8 @@ class InMemoryJobStoreTest {
         assertThat(job.getJob().getState()).isEqualTo(JobState.INITIAL);
     }
 
-    private void doNothing(final MultiTransferJob multiTransferJob) {
+    private ItemContainer doNothing(final MultiTransferJob multiTransferJob) {
+        return ItemContainer.builder().build();
     }
 
     @Test