+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/JobParameter.java
@@ -68,6 +68,9 @@ public class JobParameter {
     private boolean collectAspects;
 
     @Schema(implementation = Boolean.class)
+    private boolean integrityCheck;
+
+    @Schema(implementation = Boolean.class)
     private boolean lookupBPNs;
 
     private String callbackUrl;