+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/RegisterJob.java
@@ -80,6 +80,9 @@ public class RegisterJob {
     @Schema(description = "Flag to specify whether BPNs should be collected and resolved via the configured BPDM URL. Default is false.")
     private boolean lookupBPNs;
 
+    @Schema(description = "Specifies if integrity of the data chain is checked. Default is true.")
+    private boolean integrityCheck = true;
+
     @URL(regexp = "^(http|https).*")
     @Schema(description = "Callback url to notify requestor when job processing is finished. There are two uri variable placeholders that can be used: id and state.",
             example = "https://hostname.com/callback?id={id}&state={state}")